% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logistic}
\alias{logistic}
\title{Prevalence estimation with the logistic crosswise model}
\usage{
logistic(tables)
}
\arguments{
\item{tables}{object created by the function \code{make_tables}.}
}
\value{
\item{estimates}{matrix with theprevalence estimates for each sensitive item.}
\item{fit}{list with the observed and the fitted frequencies of the two-way
contingency tables.}
}
\description{
\code{logistic} Fits logistic models to data collected under the crosswise model.
}
\examples{
## Fit the logistic model to the items in dataset R2.
logistic(make_tables(R2))
}
