% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loglinear}
\alias{loglinear}
\title{Prevalence estimation with the loglinear crosswise model}
\usage{
loglinear(tables, model = ~., exclude = NULL, seed = 1)
}
\arguments{
\item{tables}{object created by the function \code{make_tables}.}

\item{model}{model to be fitted. The model is denoted by suffixed variables \code{Q}.}

\item{exclude}{scalar or numeric denoting the senstive item(s) to be excluded from the analysis}

\item{seed}{the seed for the start values of the loglinear parameters}
}
\value{
\item{gof}{goodness-of-fit statistics.}
\item{fitted}{multi-way contingency table with the observed and fitted frequencies.}
\item{probs}{multivariate prevalence estimates.}
\item{prev}{univariate prevalence estimates.}
\item{pars}{loglinear parameter estimates.}
}
\description{
\code{loglinear} Fits logistic models to data collected under the crosswise model.
}
\examples{
## Fit the loglinear model [Q3][Q2Q4] to the items in R2, excluding the 1st item.
loglinear(tables = make_tables(R2), model = ~ . + Q2:Q3, exclude = 1)
}
