% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_tables}
\alias{make_tables}
\title{Contigency tables for analysis with the cross-wise model}
\usage{
make_tables(df, v_order = c("1/5", "4/5"), in_answers = c("ONE", "TWO"))
}
\arguments{
\item{df}{a data frame including the subsequent pairs of individual responses to the sensitive
items, with the 1st variable pertaining to the subsample with the smallest probability of answering
the innocuous question affirmatively.}

\item{v_order}{a character vector with fractional notation of the probabilities of answering the
innocuous question within each subsample. Defaults to \code{c("1/5", "4/5")}.}

\item{in_answers}{a character vector with two strings  distinguishing between the
the reponses, with the first string corresponding to a single
affirmative response, and the second to zero or two affirmative responses. Defaults
to \code{c("ONE", "TWO")}.}
}
\value{
\item{twoway}{list with the two-way contingency tables for the subsequent items. The contingency tables
are a cross-classification of the suffixed variables \code{V}  denoting the probability of an
affirmative response to the innocuous question, and the suffixed variables \code{Q} denoting
the responses \code{both/none} and \code{one}, with the suffix denoting the sensitive item.}
\item{multiway}{a multi-way contingency table of the suffixed variables \code{V} and \code{Q}
for all items jointly.}
\item{q}{the first element of the vector \code{v_order}. To be ussed in the functions
\code{\link{logistic}} and \code{\link{loglinear}}}
}
\description{
\code{make_tables} Transforms individual responses to questions of the cross-wise
model into a contingency table format for analysis with the logistic and loglinear model.
}
\examples{
## Two-way and multiway tables for all four sensitive items.
tables_R2 <- make_tables(R2)
}
