% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PanAm}
\alias{PanAm}
\title{PanAm}
\format{
Data frame with 630 observations and the variables in columns:
\describe{
\item{1}{doping tested, subsample with probability 1/5}
\item{2}{doping tested, subsample with probability 4/5}
\item{3}{supplements, subsample with probability 1/5}
\item{4}{supplements, subsample with probability 4/5}
\item{5}{doping, subsample with probability 1/5}
\item{6}{doping, subsample with probability 4/5}
}
}
\source{
WADA.
}
\usage{
PanAm
}
\description{
Data collected during the PanAm Games of 2019 with the cross-wise model with
15-number sequences as randomizer for the innocuous questions. The three
sensitive questions involve substance use, and the responses to these
questions are collected in pairs of variables, with the 1st and 2nd variables
denoting the subsamples with respective probabilities of 1/5 and 4/5
of answering the innocuous questionan affirmatively.
}
\note{
The responses are coded "one" and "zero/two".
}
\keyword{datasets}
