% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R2}
\alias{R2}
\title{R2}
\format{
Data frame with 1211 observations and the variables in columns:
\describe{
\item{1}{drug tested, subsample with probability 1/5}
\item{2}{drug tested, subsample with probability 4/5}
\item{3}{supplements, subsample with probability 1/5}
\item{4}{supplements, subsample with probability 4/5}
\item{5}{psychoactive drugs, subsample with probability 1/5}
\item{6}{psychoactive drugs, subsample with probability 4/5}
\item{7}{doping, subsample with probability 1/5}
\item{8}{doping, subsample with probability 4/5}
}
}
\source{
WADA.
}
\usage{
R2
}
\description{
Data collected via an online forum in 2020 with the cross-wise model with
5-number sequences as randomizer for the innocuous questions. The four
sensitive questions involve substance use, and the responses to these
questions are collected in pairs of variables, with the 1st and 2nd variables
denoting the subsamples with respective probabilities of 1/5 and 4/5
of answering the innocuous questionan affirmatively.
}
\note{
The responses are coded "I've only ONE 'YES' answer" and
"I've TWO 'YES' or TWO 'NO' answers".
}
\keyword{datasets}
